Change argument name as it is not fine tuning
Change argument name as it is not fine tuning. Previous, the epoch argument was labelled as the number of epochs used in "fine tuning". However, no fine tuning is currently done in this script, so the name change reflects what the argument actually does. The number of epochs reflects the number of epochs used to train all tasks jointly.